hw4/hw4-2/requirements.txt
Cython
seaborn
scipy
PyYAML
numpy
h5py
tqdm
transformers